# specs/1-prod-todo-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Phase II Todo App API
  description: API for a production-ready full-stack Todo web application.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
  /auth/login:
    post:
      summary: Log in a user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: "User's email address"
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "refresh_token=...; HttpOnly; Secure; Path=/api/v1/auth/refresh; SameSite=Strict"
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Unauthorized
  /auth/refresh:
    post:
      summary: Refresh an access token
      tags:
        - Authentication
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Invalid refresh token
  /auth/logout:
    post:
      summary: Log out a user
      tags:
        - Authentication
      responses:
        '204':
          description: Logout successful

  /tasks:
    get:
      summary: Get a list of tasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: filter_status
          in: query
          schema:
            type: string
            enum: [all, active, completed]
        - name: filter_priority
          in: query
          schema:
            type: array
            items:
              type: string
        - name: filter_tags
          in: query
          schema:
            type: array
            items:
              type: integer
        - name: sort
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{task_id}:
    put:
      summary: Update a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted

  /tags:
    get:
      summary: Get all tags for the user
      tags:
        - Tags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a new tag
      tags:
        - Tags
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 5000
        due_date:
          type: string
          format: date
        priority:
          type: string
          enum: [Urgent, High, Medium, Low]
        status:
          type: string
          enum: [active, completed]
        position:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 5000
        due_date:
          type: string
          format: date
        priority:
          type: string
          enum: [Urgent, High, Medium, Low]
        tag_ids:
          type: array
          items:
            type: integer
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 5000
        due_date:
          type: string
          format: date
        priority:
          type: string
          enum: [Urgent, High, Medium, Low]
        status:
          type: string
          enum: [active, completed]
        position:
          type: integer
        tag_ids:
          type: array
          items:
            type: integer
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    TagCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string